# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FaceTrack API
  description: 'FaceTrack attendance system API (requirements §12). Portal: session
    cookie + CSRF header; integrations: X-API-Key.'
  version: 1.0.0
paths:
...
  /api/v1/cameras:
    get:
      tags:
      - cameras
      summary: List cameras with live health
      operationId: list_cameras
      parameters:
      - name: location_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Location Id
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/CameraRole'
          - type: 'null'
          title: Role
      - name: engine_node
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Engine Node
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
            pattern: ^-?[a-z_]+$
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_CameraOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - cameras
      summary: Add a camera
      description: The RTSP URL (with credentials) is stored AES-256 encrypted and
        never returned (FR-1, NFR-8).
      operationId: create_camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CameraCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CameraOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cameras/{camera_id}:
    get:
      tags:
      - cameras
      summary: Get Camera
      operationId: get_camera
      parameters:
      - name: camera_id
        in: path
        required: true
        schema:
          type: integer
          title: Camera Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CameraOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - cameras
      summary: Update Camera
      operationId: update_camera
      parameters:
      - name: camera_id
        in: path
        required: true
        schema:
          type: integer
          title: Camera Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CameraUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CameraOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - cameras
      summary: Delete Camera
      operationId: delete_camera
      parameters:
      - name: camera_id
        in: path
        required: true
        schema:
          type: integer
          title: Camera Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/cameras/{camera_id}/test:
    post:
      tags:
      - cameras
      summary: Test RTSP and return a snapshot
      description: FR-2. Returns 503 when the engine is unreachable.
      operationId: test_camera
      parameters:
      - name: camera_id
        in: path
        required: true
        schema:
          type: integer
          title: Camera Id
      - name: use_substream
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Use Substream
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CameraTestOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_CameraOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/CameraOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[CameraOut]
    ApiResponse_CameraTestOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          $ref: '#/components/schemas/CameraTestOut'
      type: object
      required:
      - message
      - data
      title: ApiResponse[CameraTestOut]
...
    CameraCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        location_id:
          type: integer
          title: Location Id
        role:
          $ref: '#/components/schemas/CameraRole'
        rtsp_url:
          type: string
          maxLength: 1024
          minLength: 8
          pattern: ^(rtsp|rtsps|file)://\S+$
          title: Rtsp Url
        substream_url:
          anyOf:
          - type: string
            maxLength: 1024
            pattern: ^(rtsp|rtsps|file)://\S+$
          - type: 'null'
          title: Substream Url
        engine_node:
          type: string
          maxLength: 64
          minLength: 1
          title: Engine Node
          default: node-1
        priority:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Priority
          default: 0
        operating_hours:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Operating Hours
        roi_polygon:
          anyOf:
          - items:
              items:
                type: number
              type: array
            type: array
          - type: 'null'
          title: Roi Polygon
        fps:
          anyOf:
          - type: number
            maximum: 15.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Fps
        match_threshold:
          anyOf:
          - type: number
            exclusiveMaximum: 1.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Match Threshold
        liveness_enabled:
          type: boolean
          title: Liveness Enabled
          default: false
        is_enabled:
          type: boolean
          title: Is Enabled
          default: true
      additionalProperties: false
      type: object
      required:
      - name
      - location_id
      - role
      - rtsp_url
      title: CameraCreate
...
    CameraMode:
      type: string
      enum:
      - IDLE
      - ACTIVE
      - COOLDOWN
      - PAUSED
      title: CameraMode
      description: Per-camera processing mode (requirements §10.4).
    CameraOut:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        location_id:
          type: integer
          title: Location Id
        role:
          $ref: '#/components/schemas/CameraRole'
        stream_host:
          type: string
          title: Stream Host
        has_substream:
          type: boolean
          title: Has Substream
        engine_node:
          type: string
          title: Engine Node
        priority:
          type: integer
          title: Priority
        operating_hours:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Operating Hours
        roi_polygon:
          anyOf:
          - items:
              items:
                type: number
              type: array
            type: array
          - type: 'null'
          title: Roi Polygon
        fps:
          anyOf:
          - type: number
          - type: 'null'
          title: Fps
        match_threshold:
          anyOf:
          - type: number
          - type: 'null'
          title: Match Threshold
        liveness_enabled:
          type: boolean
          title: Liveness Enabled
        is_enabled:
          type: boolean
          title: Is Enabled
        status:
          $ref: '#/components/schemas/CameraStatus'
        last_seen_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Seen At
        runtime:
          anyOf:
          - $ref: '#/components/schemas/CameraRuntimeOut'
          - type: 'null'
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - name
      - location_id
      - role
      - stream_host
      - has_substream
      - engine_node
      - priority
      - operating_hours
      - roi_polygon
      - fps
      - match_threshold
      - liveness_enabled
      - is_enabled
      - status
      - last_seen_at
      - created_at
      title: CameraOut
    CameraRole:
      type: string
      enum:
      - ENTRY
      - EXIT
      - ENTRY_EXIT
      - GENERAL
      title: CameraRole
      description: FR-1 camera roles.
    CameraRuntimeOut:
      properties:
        mode:
          $ref: '#/components/schemas/CameraMode'
        connected:
          type: boolean
          title: Connected
        fps_target:
          type: number
          title: Fps Target
        fps_actual:
          type: number
          title: Fps Actual
        lag_seconds:
          type: number
          title: Lag Seconds
        faces_today:
          type: integer
          title: Faces Today
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        updated_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Updated At
      type: object
      required:
      - mode
      - connected
      - fps_target
      - fps_actual
      - lag_seconds
      - faces_today
      - last_error
      - updated_at
      title: CameraRuntimeOut
    CameraStatus:
      type: string
      enum:
      - unknown
      - online
      - offline
      - disabled
      title: CameraStatus
      description: Connection health of a camera as last reported by its engine node
        (FR-3).
    CameraTestOut:
      properties:
        ok:
          type: boolean
          title: Ok
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        width:
          anyOf:
          - type: integer
          - type: 'null'
          title: Width
        height:
          anyOf:
          - type: integer
          - type: 'null'
          title: Height
        codec:
          anyOf:
          - type: string
          - type: 'null'
          title: Codec
        snapshot_jpeg_b64:
          anyOf:
          - type: string
          - type: 'null'
          title: Snapshot Jpeg B64
      type: object
      required:
      - ok
      - error
      - width
      - height
      - codec
      - snapshot_jpeg_b64
      title: CameraTestOut
    CameraUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Name
        location_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Location Id
        role:
          anyOf:
          - $ref: '#/components/schemas/CameraRole'
          - type: 'null'
        rtsp_url:
          anyOf:
          - type: string
            maxLength: 1024
            pattern: ^(rtsp|rtsps|file)://\S+$
          - type: 'null'
          title: Rtsp Url
        substream_url:
          anyOf:
          - type: string
            maxLength: 1024
            pattern: ^(rtsp|rtsps|file)://\S+$
          - type: 'null'
          title: Substream Url
        clear_substream:
          type: boolean
          title: Clear Substream
          default: false
        engine_node:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Engine Node
        priority:
          anyOf:
          - type: integer
            maximum: 100.0
            minimum: 0.0
          - type: 'null'
          title: Priority
        operating_hours:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Operating Hours
        roi_polygon:
          anyOf:
          - items:
              items:
                type: number
              type: array
            type: array
          - type: 'null'
          title: Roi Polygon
        clear_roi:
          type: boolean
          title: Clear Roi
          default: false
        fps:
          anyOf:
          - type: number
            maximum: 15.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Fps
        match_threshold:
          anyOf:
          - type: number
            exclusiveMaximum: 1.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Match Threshold
        liveness_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Liveness Enabled
        is_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Enabled
      additionalProperties: false
      type: object
      title: CameraUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_CameraOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        message:
          type: string
          title: Message
        data:
          items:
            $ref: '#/components/schemas/CameraOut'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - message
      - data
      - pagination
      title: PaginatedResponse[CameraOut]
...
    Pagination:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
        total:
          type: integer
          minimum: 0.0
          title: Total
      type: object
      required:
      - page
      - page_size
      - total
      title: Pagination
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError